Address [checklist](https://github.com/GMOD/Apollo3/pull/354#pullrequestreview-1864242924)
This commit should address most of the items in
https://github.com/GMOD/Apollo3/pull/354#pullrequestreview-1864242924

There are tests for most of them.

`"default.rootCredentials.password" must be a string` occurs if an entry is
expected to be string and instead is a number without quotes. The cli quotes as
appropriate.

> Make sure all commands print a sensible error message if the config file is
> not valid YAML or doesn't pass validation

Currently, the cli fails with a fairly friendly message if the yaml is
malformed, but if it is valid, it will process it anyway