Improve error handling #398
Mostly removing try/catch to let the cli fail with the full stack trace.
Replace `this.logToSterr+this.exit(1)` with `this.error`.

`this.debug` feature not used yet.

NB: Some `this.exit(0)` left since otherwise the cli hangs without
exiting (?!). To reproduce: in `add-fasta.ts` use `return` instead of
`this.exit(0)` and run

```
./test/test.py TestCLI.testAddAssemblyFromLocalFasta
```