Fix #506: Warnings do not disappear after fixing
The problem is a mismatch between `name` in CheckResult.ts and `name` in
Check.ts. In CheckResult.ts name is (was) something like
"InternalStopCodonCheck", but in Check.ts it is something like "CDSCheck". So
when calling `clearChecksForFeature` in check.service.ts, the query returns
nothing because 'CDSCheck' != 'InternalStopCodonCheck'.

This fix sets `name = "CDSCheck"` in class CDSCheck and adds a property `cause`
to CheckResult to store the cause of the failing check, e.g cause =
InternalStopCodonCheck. Also add property `causes` to class `Check` to
enumerate all the possible causes for the fail.

Note test in test.ts to be fixed: FIXME: Checks stay after invalid
operation