Improve navigatation: go to up/downstream exon, go to from table editor (#671)
* Address #634: Menu items (with icons) to go up/downstream

* This script is not needed anymore

* Fix import

* Navigate to up/downstream exon and pad

The landing coordinates fully contain the up/downstream exon plus some padding
left and right. At the moment the padding is 10% of the exon size.

* Add regions close to refseq start and end

* Implements go to feature from table editor (#672)

Double click on a row in the table editor to navigate there.

Move `navToFeatureCenter` function to `glyphUtils.ts` to share it with
"go to up/downstream exon" functionality.

Add cypress test for go-to from table editor.

* Fix lint

* Remove import from dist

---------

Co-authored-by: Garrett Stevens <stevens.garrett.j@gmail.com>